+++ master/internal/trial_test.go
@@ -4,11 +4,9 @@ import (
 	"crypto/rand"
 	"fmt"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
-	"gotest.tools/assert"
 
 	"github.com/determined-ai/determined/master/internal/mocks"
 	"github.com/determined-ai/determined/master/internal/sproto"
@@ -24,11 +22,7 @@ import (
 	"github.com/determined-ai/determined/master/pkg/searcher"
 	"github.com/determined-ai/determined/master/pkg/tasks"
 
-	"github.com/davecgh/go-spew/spew"
-
 	cproto "github.com/determined-ai/determined/master/pkg/container"
-
-	"github.com/google/uuid"
 )
 
 type (
@@ -116,7 +110,7 @@ func TestTrialMultiAlloc(t *testing.T) {
 	// Pre-allocated stage.
 	mockAlloc := func(cID cproto.ID, agentID string) sproto.Reservation {
 		alloc := &mocks.Allocation{}
-		alloc.On("Start", mock.Anything, mock.Anything, mock.Anything).Return()
+		alloc.On("Start", mock.Anything, mock.Anything, mock.Anything).Return().Times(1)
 		alloc.On("Summary").Return(sproto.ContainerSummary{
 			AllocationID: tr.req.AllocationID,
 			ID:           cID,
@@ -126,12 +120,11 @@ func TestTrialMultiAlloc(t *testing.T) {
 		return alloc
 	}
 
-	allocations := []sproto.Reservation{
+	reservations := []sproto.Reservation{
 		mockAlloc(cproto.NewID(), "agent-1"),
 		mockAlloc(cproto.NewID(), "agent-2"),
 	}
 	db.On("AddTrial", mock.Anything).Return(nil)
-	db.On("UpdateTrial", 0, model.ActiveState).Return(nil)
 	db.On("UpdateTrialRunID", 0, 1).Return(nil)
 	db.On("AddAllocation", tr.taskID, tr.req.AllocationID, mock.Anything).Return(nil)
 	db.On("StartAllocationSession", tr.req.AllocationID).Return("", nil)
@@ -141,13 +134,14 @@ func TestTrialMultiAlloc(t *testing.T) {
 		msg: sproto.ResourcesAllocated{
 			ID:           tr.req.AllocationID,
 			ResourcePool: "default",
-			Reservations: allocations,
+			Reservations: reservations,
 		},
 	}).Error())
-	require.NotEmpty(t, tr.allocations)
+	require.NotNil(t, tr.allocation)
+	require.True(t, db.AssertExpectations(t))
 
 	// Pre-ready stage.
-	for _, a := range allocations {
+	for _, a := range reservations {
 		containerStateChanged := sproto.TaskContainerStateChanged{
 			Container: cproto.Container{
 				Parent:  actor.Address{},
@@ -179,9 +173,10 @@ func TestTrialMultiAlloc(t *testing.T) {
 			containerID:  a.Summary().ID,
 		}).Error())
 	}
-	require.True(t, tr.rendezvous.ready())
+	require.True(t, tr.allocation.rendezvous.ready())
 
 	// Running stage.
+	db.On("UpdateTrial", 0, model.StoppingCompletedState).Return(nil)
 	require.NoError(t, system.Ask(self, TrialSearcherState{
 		Create: searcher.Create{RequestID: rID},
 		Op: searcher.ValidateAfter{
@@ -191,13 +186,13 @@ func TestTrialMultiAlloc(t *testing.T) {
 		Complete: true,
 		Closed:   true,
 	}).Error())
+	require.True(t, db.AssertExpectations(t))
 
 	// Terminating stage.
 	db.On("DeleteAllocationSession", tr.req.AllocationID).Return(nil)
 	db.On("CompleteAllocation", tr.req.AllocationID).Return(nil)
-	db.On("EndTasks", mock.Anything, 0).Return(nil)
 	db.On("UpdateTrial", 0, model.CompletedState).Return(nil)
-	for _, a := range allocations {
+	for _, a := range reservations {
 		containerStateChanged := sproto.TaskContainerStateChanged{
 			Container: cproto.Container{
 				Parent:  actor.Address{},
@@ -216,6 +211,7 @@ func TestTrialMultiAlloc(t *testing.T) {
 
 	require.True(t, model.TerminalStates[tr.state])
 	require.NoError(t, self.AwaitTermination())
+	require.True(t, db.AssertExpectations(t))
 }
 
 func TestTrialDelayedSearcherClose(t *testing.T) {
@@ -286,12 +282,11 @@ func TestTrialDelayedSearcherClose(t *testing.T) {
 		return alloc
 	}
 
-	allocations := []sproto.Reservation{
+	reservations := []sproto.Reservation{
 		mockAlloc(cproto.NewID(), "agent-1"),
 		mockAlloc(cproto.NewID(), "agent-2"),
 	}
 	db.On("AddTrial", mock.Anything).Return(nil)
-	db.On("UpdateTrial", 0, model.ActiveState).Return(nil)
 	db.On("UpdateTrialRunID", 0, 1).Return(nil)
 	db.On("AddAllocation", tr.taskID, tr.req.AllocationID, mock.Anything).Return(nil)
 	db.On("StartAllocationSession", tr.req.AllocationID).Return("", nil)
@@ -301,13 +296,14 @@ func TestTrialDelayedSearcherClose(t *testing.T) {
 		msg: sproto.ResourcesAllocated{
 			ID:           tr.req.AllocationID,
 			ResourcePool: "default",
-			Reservations: allocations,
+			Reservations: reservations,
 		},
 	}).Error())
-	require.NotEmpty(t, tr.allocations)
+	require.NotNil(t, tr.allocation)
+	require.True(t, db.AssertExpectations(t))
 
 	// Pre-ready stage.
-	for _, a := range allocations {
+	for _, a := range reservations {
 		containerStateChanged := sproto.TaskContainerStateChanged{
 			Container: cproto.Container{
 				Parent:  actor.Address{},
@@ -339,7 +335,7 @@ func TestTrialDelayedSearcherClose(t *testing.T) {
 			containerID:  a.Summary().ID,
 		}).Error())
 	}
-	require.True(t, tr.rendezvous.ready())
+	require.True(t, tr.allocation.rendezvous.ready())
 
 	// Running stage.
 	require.NoError(t, system.Ask(self, TrialSearcherState{
@@ -356,8 +352,8 @@ func TestTrialDelayedSearcherClose(t *testing.T) {
 	// Terminating stage.
 	db.On("DeleteAllocationSession", tr.req.AllocationID).Return(nil)
 	db.On("CompleteAllocation", tr.req.AllocationID).Return(nil)
-	db.On("UpdateTrialRunIDAndRestarts", 0, 1, 0).Return(nil)
-	for _, a := range allocations {
+	db.On("UpdateTrialRunID", 0, 1).Return(nil)
+	for _, a := range reservations {
 		containerStateChanged := sproto.TaskContainerStateChanged{
 			Container: cproto.Container{
 				Parent:  actor.Address{},
@@ -373,9 +369,10 @@ func TestTrialDelayedSearcherClose(t *testing.T) {
 		}
 		require.NoError(t, system.Ask(self, containerStateChanged).Error())
 	}
+	require.True(t, db.AssertExpectations(t))
 
 	// Later searcher decides it's done.
-	db.On("EndTasks", mock.Anything, 0).Return(nil)
+	db.On("UpdateTrial", 0, model.StoppingCompletedState).Return(nil)
 	db.On("UpdateTrial", 0, model.CompletedState).Return(nil)
 	require.NoError(t, system.Ask(self, TrialSearcherState{
 		Create: searcher.Create{RequestID: rID},
@@ -389,6 +386,7 @@ func TestTrialDelayedSearcherClose(t *testing.T) {
 
 	require.True(t, model.TerminalStates[tr.state])
 	require.NoError(t, self.AwaitTermination())
+	require.True(t, db.AssertExpectations(t))
 }
 
 func TestTrialRestarts(t *testing.T) {
@@ -458,7 +456,6 @@ func TestTrialRestarts(t *testing.T) {
 		})
 		alloc.On("Kill", mock.Anything).Return()
 		db.On("AddTrial", mock.Anything).Return(nil)
-		db.On("UpdateTrial", 0, model.ActiveState).Return(nil)
 		db.On("UpdateTrialRunID", 0, i+1).Return(nil)
 		db.On("AddAllocation", tr.taskID, tr.req.AllocationID, mock.Anything).Return(nil)
 		db.On("StartAllocationSession", tr.req.AllocationID).Return("", nil)
@@ -471,7 +468,8 @@ func TestTrialRestarts(t *testing.T) {
 				Reservations: []sproto.Reservation{alloc},
 			},
 		}).Error())
-		require.NotEmpty(t, tr.allocations)
+		require.NotNil(t, tr.allocation)
+		require.True(t, db.AssertExpectations(t))
 
 		// Pre-ready stage.
 		containerStateChanged := sproto.TaskContainerStateChanged{
@@ -504,7 +502,7 @@ func TestTrialRestarts(t *testing.T) {
 			allocationID: tr.req.AllocationID,
 			containerID:  cID,
 		}).Error())
-		require.True(t, tr.rendezvous.ready())
+		require.True(t, tr.allocation.rendezvous.ready())
 
 		// Running stage.
 
@@ -513,7 +511,6 @@ func TestTrialRestarts(t *testing.T) {
 		db.On("CompleteAllocation", tr.req.AllocationID).Return(nil)
 		db.On("UpdateTrialRestarts", 0, i+1).Return(nil)
 		if i == tr.config.MaxRestarts() {
-			db.On("EndTasks", mock.Anything, 0).Return(nil)
 			db.On("UpdateTrial", 0, model.ErrorState).Return(nil)
 		}
 		containerStateChanged.Container.State = cproto.Terminated
@@ -527,268 +524,8 @@ func TestTrialRestarts(t *testing.T) {
 			}},
 		}
 		require.NoError(t, system.Ask(self, containerStateChanged).Error())
+		require.True(t, db.AssertExpectations(t))
 	}
 	require.True(t, model.TerminalStates[tr.state])
 	require.NoError(t, self.AwaitTermination())
-}
-
-func TestRendezvous(t *testing.T) {
-	const operations = 4
-	type testCase struct {
-		name  string
-		order []int
-	}
-
-	runTestCase := func(t *testing.T, tc testCase) {
-		t.Run(tc.name, func(t *testing.T) {
-			// "task" with ranks is started.
-			t1 := model.NewAllocationID(uuid.New().String())
-			c1, c2 := cproto.NewID(), cproto.NewID()
-			ranks := map[cproto.ID]int{c1: 0, c2: 1}
-			r := newRendezvous(t1, ranks)
-
-			assert.Equal(t, r.rank(c1), 0)
-			assert.Equal(t, r.rank(c2), 1)
-
-			var ws []rendezvousWatcher
-			watch := func(cID cproto.ID) func() {
-				return func() {
-					w, err := r.watch(t1, cID)
-					assert.NilError(t, err, cID)
-					ws = append(ws, w)
-				}
-			}
-
-			startContainer := func(cID cproto.ID) func() {
-				return func() {
-					r.containerStarted(cID, addressesFromContainerID(cID))
-				}
-			}
-
-			ops := []func(){
-				watch(c1),
-				watch(c2),
-				startContainer(c1),
-				startContainer(c2),
-			}
-			for _, i := range tc.order {
-				assert.Check(t, !r.ready())
-				ops[i]()
-			}
-			assert.Check(t, r.ready())
-
-			rendezvousArrived := func(w rendezvousWatcher) {
-				select {
-				case resp := <-w.C:
-					assert.NilError(t, resp.err)
-					assert.Equal(t, len(resp.info.Addresses), 2)
-				default:
-					t.Fatal("expected rendezvous on first watcher but found none")
-				}
-			}
-			for _, w := range ws {
-				rendezvousArrived(w)
-			}
-
-			r.unwatch(c1)
-			r.unwatch(c2)
-		})
-	}
-
-	for _, ordering := range orderings(operations) {
-		runTestCase(t, testCase{
-			name:  spew.Sdump(ordering),
-			order: ordering,
-		})
-	}
-}
-
-func TestRendezvousUninitialized(t *testing.T) {
-	// Initialize a nil rendezvous
-	var r *rendezvous
-
-	// All API-connected methods (so ones a user could call) should not panic the actor.
-	tID := model.NewAllocationID(uuid.New().String())
-	cID := cproto.NewID()
-	_, err := r.watch(tID, cID)
-	assert.ErrorContains(t, err, "watch rendezvous not valid without active task")
-	r.unwatch(cID)
-}
-
-func TestRendezvousValidation(t *testing.T) {
-	t1 := model.NewAllocationID(uuid.New().String())
-	c1 := cproto.NewID()
-	r := newRendezvous(t1, map[cproto.ID]int{
-		c1: 0,
-	})
-
-	_, err := r.watch(t1, cproto.NewID())
-	assert.ErrorContains(t, err, "stale container")
-
-	_, err = r.watch(t1, c1)
-	assert.NilError(t, err)
-
-	_, err = r.watch(t1, c1)
-	assert.ErrorContains(t, err, "rendezvous request from already connected container")
-}
-
-func TestTerminationInRendezvous(t *testing.T) {
-	t1 := model.NewAllocationID(uuid.New().String())
-	c1, c2 := cproto.NewID(), cproto.NewID()
-	ranks := map[cproto.ID]int{c1: 0, c2: 1}
-	r := newRendezvous(t1, ranks)
-
-	r.containerStarted(c1, addressesFromContainerID(c1))
-	_, err := r.watch(t1, c1)
-	assert.NilError(t, err)
-	r.containerTerminated(c1)
-
-	r.containerStarted(c2, addressesFromContainerID(c2))
-	_, err = r.watch(t1, c2)
-	assert.NilError(t, err)
-
-	assert.Check(t, !r.ready())
-}
-
-func TestUnwatchInRendezvous(t *testing.T) {
-	t1 := model.NewAllocationID(uuid.New().String())
-	c1, c2 := cproto.NewID(), cproto.NewID()
-	ranks := map[cproto.ID]int{c1: 0, c2: 1}
-	r := newRendezvous(t1, ranks)
-
-	r.containerStarted(c1, addressesFromContainerID(c1))
-	_, err := r.watch(t1, c1)
-	assert.NilError(t, err)
-	r.unwatch(c1)
-
-	r.containerStarted(c2, addressesFromContainerID(c2))
-	_, err = r.watch(t1, c2)
-	assert.NilError(t, err)
-
-	assert.Check(t, !r.ready())
-}
-
-func TestRendezvousTimeout(t *testing.T) {
-	rendezvousTimeoutDuration = 0
-
-	t1 := model.NewAllocationID(uuid.New().String())
-	c1, c2 := cproto.NewID(), cproto.NewID()
-	ranks := map[cproto.ID]int{c1: 0, c2: 1}
-	r := newRendezvous(t1, ranks)
-
-	_, err := r.watch(t1, c1)
-	assert.NilError(t, err)
-	r.containerStarted(c1, addressesFromContainerID(c1))
-
-	time.Sleep(-1)
-	assert.ErrorContains(t, r.checkTimeout(t1), "some containers are taking a long time")
-}
-
-func addressesFromContainerID(cID cproto.ID) []cproto.Address {
-	return []cproto.Address{
-		{
-			ContainerIP:   "172.0.1.2",
-			ContainerPort: 1734,
-			HostIP:        fmt.Sprintf("%s.somehost.io", cID),
-			HostPort:      1734,
-		},
-	}
-}
-
-func TestPreemption(t *testing.T) {
-	// Initialize a nil preemption.
-	var p *preemption
-
-	// Watch nil should not panic and return an error.
-	id := uuid.New()
-	_, err := p.watch(model.NewAllocationID(uuid.New().String()), id)
-	assert.ErrorContains(t, err, "no preemption status")
-
-	// All method on nil should not panic.
-	p.unwatch(id)
-	p.preempt()
-	p.close()
-
-	// "task" is allocated.
-	t1 := model.NewAllocationID(uuid.New().String())
-	p = newPreemption(t1)
-
-	// real watcher connects
-	id = uuid.New()
-	w, err := p.watch(t1, id)
-	assert.NilError(t, err)
-
-	// should immediately receive initial status.
-	select {
-	case <-w.C:
-		t.Fatal("received preemption but should not have")
-	default:
-	}
-
-	// on preemption, it should also receive status.
-	p.preempt()
-
-	// should receive updated preemption status.
-	select {
-	case <-w.C:
-	default:
-		t.Fatal("did not receive preemption")
-	}
-
-	// preempted preemption unwatching should work.
-	p.unwatch(id)
-
-	// new post-preemption watch connects
-	id = uuid.New()
-	w, err = p.watch(t1, id)
-	assert.NilError(t, err)
-
-	// should immediately receive initial status and initial status should be preemption.
-	select {
-	case <-w.C:
-	default:
-		t.Fatal("preemptionWatcher.C was empty channel (should come with initial status when preempted)")
-	}
-
-	// preempted preemption unwatching should work.
-	p.unwatch(id)
-}
-
-// orderings returns all orders for n operations.
-func orderings(n int) [][]int {
-	var xs []int
-	for i := 0; i < n; i++ {
-		xs = append(xs, i)
-	}
-	return permutations(xs)
-}
-
-// https://stackoverflow.com/questions/30226438/generate-all-permutations-in-go
-func permutations(arr []int) [][]int {
-	var helper func([]int, int)
-	res := [][]int{}
-
-	helper = func(arr []int, n int) {
-		if n == 1 {
-			tmp := make([]int, len(arr))
-			copy(tmp, arr)
-			res = append(res, tmp)
-		} else {
-			for i := 0; i < n; i++ {
-				helper(arr, n-1)
-				if n%2 == 1 {
-					tmp := arr[i]
-					arr[i] = arr[n-1]
-					arr[n-1] = tmp
-				} else {
-					tmp := arr[0]
-					arr[0] = arr[n-1]
-					arr[n-1] = tmp
-				}
-			}
-		}
-	}
-
-	helper(arr, len(arr))
-	return res
 }
